+++ Samples~/Demo/Scripts/AdvancedPopup_GlobalManager_Demo.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-public class AdvancedPopup_GlobalManager_Demo : MonoBehaviour
-{
-    private void Start()
-    {
-        AdvancedPopupSystem.ForceShow<SmoothScaleDisplay, ImmediatelyScaleDisplay>("Shop");
-    }
-}