+++ Scripts/Runtime/Models/Popups/AdvancedPopup.cs
@@ -36,7 +36,31 @@ namespace AdvancedPS.Core.Popup
             _closeButton.onClick.RemoveListener(OnCloseButtonPress);
             _subscribed = false;
         }
-    
+
+        public override async Task Show(bool deepShow = false, CancellationToken cancellationToken = default)
+        {
+            if (cancellationToken == default)
+                cancellationToken = UpdateCancellationTokenSource();
+        
+            if (!_subscribed)
+                Subscribe();
+        
+            List<Task> tasks = new List<Task>();
+        
+            IAdvancedPopupDisplay popupDisplay = base.CachedDisplay;
+            tasks.Add(popupDisplay.Show(RootTransform, cancellationToken));
+
+            if (deepShow)
+            {
+                foreach (IAdvancedPopup popup in DeepPopups)
+                {
+                    tasks.Add(popup.Show(true, cancellationToken));
+                }
+            }
+            if (tasks.Count > 0)
+                await Task.WhenAll(tasks);
+        }
+
         public override async Task Show<T>(bool deepShow = false, CancellationToken cancellationToken = default)
         {
             if (cancellationToken == default)
@@ -54,7 +78,7 @@ namespace AdvancedPS.Core.Popup
             {
                 foreach (IAdvancedPopup popup in DeepPopups)
                 {
-                    tasks.Add(popup.Show<T>(deepShow, cancellationToken));
+                    tasks.Add(popup.Show<T>(true, cancellationToken));
                 }
             }
             if (tasks.Count > 0)
@@ -78,11 +102,35 @@ namespace AdvancedPS.Core.Popup
             {
                 foreach (IAdvancedPopup popup in DeepPopups)
                 {
-                    tasks.Add(popup.Show<J>(deepShow, cancellationToken));
+                    tasks.Add(popup.Show<J>(true, cancellationToken));
                 }
             }
             if (tasks.Count > 0)
                 await Task.WhenAll(tasks);
+        }
+
+        public override async Task Hide(bool deepHide = false, CancellationToken cancellationToken = default)
+        {
+            if (cancellationToken == default)
+                cancellationToken = UpdateCancellationTokenSource();
+        
+            List<Task> tasks = new List<Task>();
+        
+            IAdvancedPopupDisplay popupDisplay = base.CachedDisplay;
+            tasks.Add(popupDisplay.Hide(RootTransform, cancellationToken));
+
+            if (deepHide)
+            {
+                foreach (IAdvancedPopup popup in DeepPopups)
+                {
+                    tasks.Add(popup.Hide(true, cancellationToken));
+                }
+            }
+            if (tasks.Count > 0)
+                await Task.WhenAll(tasks);
+        
+            if (_subscribed)
+                Unsubscribe();
         }
 
         public override async Task Hide<T>(bool deepHide = false, CancellationToken cancellationToken = default)
@@ -99,7 +147,7 @@ namespace AdvancedPS.Core.Popup
             {
                 foreach (IAdvancedPopup popup in DeepPopups)
                 {
-                    tasks.Add(popup.Hide<T>(deepHide, cancellationToken));
+                    tasks.Add(popup.Hide<T>(true, cancellationToken));
                 }
             }
             if (tasks.Count > 0)
@@ -123,7 +171,7 @@ namespace AdvancedPS.Core.Popup
             {
                 foreach (IAdvancedPopup popup in DeepPopups)
                 {
-                    tasks.Add(popup.Hide<J>(deepHide, cancellationToken));
+                    tasks.Add(popup.Hide<J>(true, cancellationToken));
                 }
             }
             if (tasks.Count > 0)