+++ Scripts/Runtime/Models/Popups/AdvancedPopupInstantiate.cs
@@ -5,6 +5,11 @@ namespace AdvancedPS.Core.Popup
 {
     public class AdvancedPopupInstantiate : IAdvancedPopup
     {
+        public override async Task Show(bool deepShow = false, CancellationToken cancellationToken = default)
+        {
+            
+        }
+
         public override async Task Show<T>(bool deepShow = false, CancellationToken cancellationToken = default)
         {
         
@@ -13,6 +18,11 @@ namespace AdvancedPS.Core.Popup
         public override async Task Show<T, J>(bool deepShow = false, CancellationToken cancellationToken = default)
         {
         
+        }
+
+        public override async Task Hide(bool deepHide = false, CancellationToken cancellationToken = default)
+        {
+            
         }
 
         public override async Task Hide<T>(bool deepHide = false, CancellationToken cancellationToken = default)