+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
-## [1.0.3] - 2022-11-10
+## [1.0.4] - 2023-4-21
+### Remove
+- Show/hide popups by name
+- Old demo's
+- ### Added
+- Caching IAdvancedPopupDisplay for popups
+- Show/hide popup without generic methods by cached IAdvancedPopupDisplay type
+### Rework
+- Tooltip's $ summary's updated
+
+## [1.0.3] - 2023-2-24
 ### Rework
 - Improvement and polishing of systems after combat use in projects
 
-## [1.0.2] - 2022-11-10
+## [1.0.2] - 2022-11-14
 ### Fixed
 - Fixed layers~